Added the searchbar component. Cleaned up the sidebar component.
