+++ src/Searchbar/Searchbar.scss
@@ -1,0 +1,7 @@
+.highlighted {
+  color: #3baae1;
+}
+
+.ant-select-item-option-content {
+  white-space: normal;
+}
