+++ src/Sidebar/Sidebar.scss
@@ -5,18 +5,12 @@
   height: 100%;
 }
 
-.sidebar-searchbar {
-  // font: 400 11px system-ui;
+.searchbar {
   width: 100%;
   font-weight: 400;
   font-size: 20px;
 }
 
-.highlighted {
-  color: #3baae1;
-}
-
-.ant-select-item-option-content,
 .ant-card-head-title {
   white-space: normal;
 }
